requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
python-dotenv
gunicorn
langchain-chroma